+++ saleor/graphql/shipping/bulk_mutations/__init__.py
@@ -1,0 +1,4 @@
+from .shipping_price_bulk_delete import ShippingPriceBulkDelete
+from .shipping_zone_bulk_delete import ShippingZoneBulkDelete
+
+__all__ = ["ShippingPriceBulkDelete", "ShippingZoneBulkDelete"]
