+++ saleor/graphql/shipping/mutations/__init__.py
@@ -1,0 +1,23 @@
+from .shipping_method_channel_listing_update import ShippingMethodChannelListing
+from .shipping_price_create import ShippingPriceCreate
+from .shipping_price_delete import ShippingPriceDelete
+from .shipping_price_exclude_products import ShippingPriceExcludeProducts
+from .shipping_price_remove_product_from_exclude import (
+    ShippingPriceRemoveProductFromExclude,
+)
+from .shipping_price_update import ShippingPriceUpdate
+from .shipping_zone_create import ShippingZoneCreate
+from .shipping_zone_delete import ShippingZoneDelete
+from .shipping_zone_update import ShippingZoneUpdate
+
+__all__ = [
+    "ShippingMethodChannelListing",
+    "ShippingPriceCreate",
+    "ShippingPriceDelete",
+    "ShippingPriceExcludeProducts",
+    "ShippingPriceRemoveProductFromExclude",
+    "ShippingPriceUpdate",
+    "ShippingZoneCreate",
+    "ShippingZoneDelete",
+    "ShippingZoneUpdate",
+]
