+++ saleor/graphql/shipping/tests/mutations/test_shipping_zone_create.py
@@ -205,7 +205,7 @@ TEST_COUNTRIES_LIST = ["DZ", "AX", "BY"]
 
 
 @mock.patch(
-    "saleor.graphql.shipping.mutations.shippings.get_countries_without_shipping_zone",
+    "saleor.graphql.shipping.mutations.base." "get_countries_without_shipping_zone",
     return_value=TEST_COUNTRIES_LIST,
 )
 def test_create_default_shipping_zone(
